BettIR moved out and added vehicle mod
BettIR was double in the modpack and its better to have it outside the mod due to alot of mods is looking for it due to compatability.

added in the GMV which is the closest we have to the new Airborne vehicle. the mod owner allows for repacking as long as we give credit. it lacks mod key since its not been updated since 2023 so this should also solve this issue.,